finish /create

+++ springboot/src/main/java/online/exam/datacenter/VO/ExamResponse.java
@@ -9,6 +9,6 @@ import lombok.ToString;
 @ToString
 public class ExamResponse {
     private String status;
-    private long examID;
+    private int examID;
     private String password;
 }
